calloc

+++ ft_calloc.c
@@ -6,14 +6,14 @@
 /*   By: ouamarko <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/21 17:43:51 by ouamarko          #+#    #+#             */
-/*   Updated: 2025/04/21 19:21:43 by ouamarko         ###   ########.fr       */
+/*   Updated: 2025/04/22 17:08:42 by ouamarko         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
-void *ft_calloc(size_t nmemb, size_t size)
+void	*ft_calloc(size_t nmemb, size_t size)
 {
-	unsigned int	i;
+	size_t	i;
 	char	*tab;
 
 	i = 0;
